requirements.txt
streamlit>=1.33.0
langchain>=0.2.11
langchain-core>=0.2.35
langchain-text-splitters>=0.2.2
langchain-ollama>=0.1.0
langchain-community>=0.2.11
langchain-chroma>=0.1.2
chromadb>=0.5.5
pdfplumber>=0.10.3
pypdf>=4.2.0
python-dotenv>=1.0.1
langchain-google-genai>=1.0.6
langchain-openai>=0.1.16
google-generativeai>=0.8.3